Moved js files to Src folder and installed npm slash-commands
